Move Measurement to measurement package with the other models
